Header: Added group/ungroup key-combos

+++ pv/view/header.cpp
@@ -299,16 +299,16 @@ void Header::keyPressEvent(QKeyEvent *e)
 {
 	assert(e);
 
-	switch (e->key())
-	{
-	case Qt::Key_Delete:
+	if (e->key() == Qt::Key_Delete)
 	{
 		for (const shared_ptr<RowItem> r : _view)
 			if (r->selected())
 				r->delete_pressed();
-		break;
 	}
-	}
+	else if (e->key() == Qt::Key_G && e->modifiers() == Qt::ControlModifier)
+		on_group();
+	else if (e->key() == Qt::Key_U && e->modifiers() == Qt::ControlModifier)
+		on_ungroup();
 }
 
 void Header::on_signals_moved()
